feat(payment): move payment dto to folder
